% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDatavisu.R
\name{AllDatavisu}
\alias{AllDatavisu}
\title{Complete data visualization}
\usage{
AllDatavisu(
  object2,
  redmet = "PCA",
  cmet = "svm",
  ax = c(1, 2),
  Title = "Cellular map",
  Condition = "cond",
  highpr = FALSE,
  proteins = c("PKN2", "GRB2", "SHC1", "EGFR"),
  vect = FALSE,
  same_size = FALSE,
  sz = 1,
  unknow = FALSE,
  expor.png = FALSE,
  expor.pdf = FALSE,
  yourseed = 500
)
}
\arguments{
\item{object2}{A \code{\link{MSnSet}} object}

\item{redmet}{The reduction method from the pRoloc package : "PCA", "t-SNE", "MDS", "nipals", "lda", "kpca".}

\item{cmet}{the name of the column of your data which contains the markers, if it contains "unknown" assignment, please precise unknow = TRUE}

\item{ax}{A numeric vector of length two, you can choose on which axes you want to see the plot.
(depend of the number of fraction of the data)}

\item{Title}{The title of the plot, use only when vect = TRUE}

\item{Condition}{A character argument, it is the name of the column of your data which contains the condition of the experiment (control or treated).
You can have replicates.
If you have no dynamic experiment, don't bother with this parameter.}

\item{highpr}{A logical argument to highlight or not specific protein}

\item{proteins}{A character vector containing the proteins you want to highlight}

\item{vect}{A logical argument to use when you have dynamic experiment.
If TRUE, allows you to see the proteins movement between the two conditions with vectors.
You also have to choose the proteins you want to see with the proteins argument.}

\item{same_size}{A logical argument to tell if you want all the proteins having the same size on the plot.
if FALSE, the size is the score of the clustering algorithm (also if unknow = FALSE).}

\item{sz}{An integer for the size you want (if same_size = TRUE)}

\item{unknow}{A logical argument to tell if the marker column contain "unknown" features}

\item{expor.png}{A logical argument to export the figure in a png file}

\item{expor.pdf}{A logical argument to export the figure in a pdf file}

\item{yourseed}{An integer for the t-SNE algorithm in order to having same plot if there is several}
}
\value{
A figure showing your data, depending of your chosen parameters
}
\description{
\code{AllDatavisu} allows you to visualize your data with many options.
You can choose the reduction method (PCA, t-SNE, ...), the clustering method, export the plot as a png or pdf file.
The function adapts according to the data, so if you have replicates or two different experiment conditions you can visualize it.
This is very useful when you work with dynamic organellar map.
With this type of data you can visualize the movement of protein between the two conditions with a vector.
}
\examples{
library(pRolocdata)
data(tan2009r1)
AllDatavisu(datavisupca(tan2009r1, method ="knn", sh.gr = FALSE), redmet = "t-SNE", cmet = "knn")
}
\seealso{
\code{\link{plot2D}} and \code{\link{svmOptimisation}} from pRoloc package for more details
}
