% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllDatavisuInt.R
\name{AllDatavisuInt}
\alias{AllDatavisuInt}
\title{Interactive (or not) data visualization}
\usage{
AllDatavisuInt(
  object2,
  redmet = "PCA",
  cmet = "svm",
  ax = c(1, 2),
  Interact = FALSE,
  Title = "Cellular map",
  name_cond = "cond",
  Mean_point = FALSE,
  highpr = FALSE,
  proteins = c("PKN2", "GRB2", "SHC1", "EGFR"),
  vect = FALSE,
  unknow = FALSE,
  highcond = FALSE,
  Condition = 1,
  expor.png = FALSE,
  expor.pdf = FALSE,
  yourseed = 500,
  Source = "AA"
)
}
\arguments{
\item{object2}{A \code{\link{MSnSet}} object}

\item{redmet}{The reduction method from the pRoloc package : "PCA", "t-SNE", "MDS", "nipals", "lda", "kpca".}

\item{cmet}{the name of the column of your data which contains the markers, if it contains "unknown" assignment, please precise unknow = TRUE}

\item{ax}{A numeric vector of length two, you can choose on which axes you want to see the plot.
(depend of the number of fraction of the data)}

\item{Interact}{A logical argument to tell if you want interactive plot}

\item{Title}{The title of the plot, use only when vect = TRUE}

\item{name_cond}{A character argument, it is the name of the column of your data which contains the condition of the experiment (control or treated).
You can have replicates.
If you have no dynamic experiment, don't bother with this parameter.}

\item{Mean_point}{A logical argument to tell if you want to print the mean point of each cluster on the plot}

\item{highpr}{A logical argument to highlight or not specific protein}

\item{proteins}{A character vector containing the proteins you want to highlight}

\item{vect}{A logical argument to use when you have dynamic experiment.
If TRUE, allows you to see the proteins movement between the two conditions with vectors.
You also have to choose the proteins you want to see with the proteins argument.}

\item{unknow}{A logical argument to tell if the marker column contain "unknown" features}

\item{highcond}{A logical argument if you want tu highlight specific condition (replicates / control and treated)}

\item{Condition}{A numeric argument to tell which condition you want to highlight}

\item{expor.png}{A logical argument to export the figure in a png file}

\item{expor.pdf}{A logical argument to export the figure in a pdf file}

\item{yourseed}{An integer for the t-SNE algorithm in order to having same plot if there is several}

\item{Source}{A character specifying the source of the plotly graph (use when Interact = TRUE)}
}
\value{
A figure showing your data, depending of your chosen parameters
}
\description{
\code{AllDatavisuInt} allows you to visualize your data with many options.
You can choose the reduction method (PCA, t-SNE, ...), the clustering method, export the plot as a png or pdf file.
The function adapts according to the data, so if you have replicates or two different experiment conditions you can visualize it.
This is very useful when you work with dynamic organellar map.
With this type of data you can visualize the movement of protein between the two conditions with a vector.
The two many differences between \code{AllDatavisuInt} and \code{\link{AllDatavisuInt}}, are that you can visualize your
interactively with plotly, and it shows only one graph per figure.
This function was mainly created for shiny purpose. (for example run \code{rnShinyVisualization})
}
\examples{
library(pRolocdata)
data(tan2009r1)
AllDatavisuInt(datavisupca(tan2009r1, method ="knn", sh.gr = FALSE), cmet = "knn", Interact = TRUE)
}
\seealso{
\code{\link{plot2D}}, \code{\link{ggplotly}} and \code{\link{svmOptimisation}} from pRoloc package for more details
}
