% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alldyn}
\alias{alldyn}
\title{Synthetic data of dynamic organellar map, from Borner and al. 2016}
\format{
A large MSnSet
}
\usage{
alldyn

alldyn_two

alldyn_mean
}
\description{
A large \code{\link{MSnSet}} containing the data from Borner and al. 2016.
 It contains 2235 proteins, with five fractions, three replicates for two conditions.
 The data were normalized (see Borner and al. 2016 for more details)
 The best way to access its content is to use the functions :
 \code{\link{fData}} and \code{\link{exprs}},
 For alldyn : exprs show a tab of 5 columns (the five fractions) and 2235*6 rows (three replicates times two conditions).
 For alldyn_two : exprs show a tab of 15 columns (3 replicates times five fractions) and 2235*2 rows (two conditions).
 For alldyn_mean : exprs show a tab of 5 columns (the three replicates were averaged) and 2235*2 rows (two conditions).
 fData always show a tab of 4 columns containing the gene name, the markers of each protein
 (their cellular location), a columns named protcond_name to differentiate proteins between the replicates and the conditions
 and a column named cond to differentiate replicates and the conditions.
 Here \code{\link{pData}} isn't useful
}
\keyword{datasets}
