% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Borner2016}
\alias{Borner2016}
\title{Data from Borner and al. 2016, preotein cellular localization}
\format{
A large list of 12 elements
}
\usage{
data(Borner2016)
}
\description{
A large list containing the data from Borner and al. 2016.
 It contains two differents experiment, both using SILAC.
 The first is "static", with 6 replicates of 5 fractions.
 The second is dynamic, in the sense that the samples are separated with control and treated by EGF.
 The control is made of three replicates of 5 fractions, as the EGF treated.
}
\keyword{datasets}
