% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxProfile.R
\name{BoxProfile}
\alias{BoxProfile}
\title{Box plot profile for dynamic organellar map}
\usage{
BoxProfile(obj, prot = "GRB2", Condition = "cond")
}
\arguments{
\item{obj}{A \code{\link{MSnSet}} object}

\item{prot}{A character vector which contains the protein(s) that you want to see the profiles between the two conditions}

\item{Condition}{The name of the column of your data which contains the condition for each protein (control or treated)}
}
\value{
a list containing the plot and the summary of the anova analysis
}
\description{
\code{BoxProfile} plot a box plot of specific proteins between two condition (control vs treated).
Perform also an anova analysis on the protein selected.
The box are only useful when there are replicates.
This allow to visualize quickly if the protein profile changed between the two conditions.
Use only when you have dataset with two different conditions.
}
