% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPAClassification.R
\name{CPAClassification}
\alias{CPAClassification}
\title{CPA classification}
\usage{
CPAClassification(
  object,
  scores = c("prediction", "all", "none"),
  fcol = "markers",
  method = 3,
  ...
)
}
\arguments{
\item{object}{An instance of class \code{"\linkS4class{MSnSet}"}.}

\item{scores}{One of \code{"prediction"}, \code{"all"} or
\code{"none"} to report the score for the predicted class
only, for all classes or none.}

\item{fcol}{The feature meta-data containing marker definitions.
Default is \code{markers}.}

\item{method}{An integer (1, 2 or 3) specifying wich Barzilai-Borwein steplength to use; default is 3.
For more details see documentation from \code{\link{spg}} function from package \code{BB}.}

\item{...}{Additional parameters passed to \code{\link{spg}} from
package \code{BB}.}
}
\value{
An instance of class \code{"\linkS4class{MSnSet}"} with
    \code{CPA} and \code{CPA.scores} feature variables storing the
    classification results and scores respectively.
}
\description{
Classification using constrained proportionate assignment method
}
\details{
CPA was originally described by Michel Jadot et al. in 'Accounting for protein subcellular localization:
         A compartmental map of the rat liver proteome.', 2017. The goal of this method is to assign
         probabilities to each protein of belonging to an organelle. This way, we can potentially assign multiple
         organelles to proteins (a protain may have multiple locations).
         For this, it uses the mean profile of each organelle of the train dataset and then find the best coefficients
         to each profile  to obtain the profile from a given protein. The coefficients are constrained to sum to one and
         bounded between 0 and 1 (so we can interpret them as probabilities). To solve this constrained optimization
         problem, we use \code{spg} function from \code{BB} package, as in the article.
         However, in this function, any treatments on the mean profiles are done, unlike in the article. So be sure
         of your reference proteins and use normalized data.
}
\examples{
library(pRolocExtra)
data(tan2009r1)
res <- CPAClassification(tan2009r1)
getPredictions(res, fcol = "CPA")
getPredictions(res, fcol = "CPA", t = 0.75)
plot2D(res, fcol = "CPA")
}
