% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datavisupca.R
\name{datavisupca}
\alias{datavisupca}
\title{PCA visualization of your \code{\link{MSnSet}} data}
\usage{
datavisupca(
  object,
  mfcol = "markers",
  method = "svm",
  ax = c(1, 2),
  sh.gr = TRUE,
  tm = 5,
  cv = 5
)
}
\arguments{
\item{object}{A \code{\link{MSnSet}} object}

\item{mfcol}{The name of the column which contains the markers from your data}

\item{method}{The clustering method, available : svm, knn, perTurbo, nnet (for neural network), rf (for random forest) and naiveBayes}

\item{ax}{A numeric vector of length 2, the axes on which you want to see the PCA plot (depend of the number of fraction of the data)}

\item{sh.gr}{A logical argument, to show or not the plot.
if FALSE, return only a \code{\link{MSnSet}} object : your data + the clustering information}

\item{tm}{An integer corresponding to the times parameter of clustering optimization functions from pRoloc package}

\item{cv}{An integer corresponding to the cross validation parameter of clustering optimization functions from pRoloc package}
}
\value{
A list containing the two PCA plots on the same figure (clustered and not clustered) and the clustered MSnSet object if sh.gr = TRUE
else, it return only the clustered \code{\link{MSnSet}} object
}
\description{
\code{datavisupca} allows you to visualize the PCA plot of your data, clustered and not clustered on the same figure.
You can also choose to not see the plot and only get back your clustered data.
The clustering method available are from the pRoloc package (Laurent Gatto and al.).
}
\examples{
library(pRolocdata)
data(tan2009r1)
datavisupca(tan2009r1)
}
\seealso{
\code{\link{svmOptimisation}} from Roloc package and \code{\link{fviz_pca_ind}} from factoextra package
}
