% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorClustering.R
\name{ErrorClustering}
\alias{ErrorClustering}
\title{Get the approximate error of a clustering algorithm}
\usage{
ErrorClustering(data, n, cmet = "knn", train_size = 0.8)
}
\arguments{
\item{data}{A \code{\link{MSnSet}} object}

\item{n}{An integer which correspond to the number of loop}

\item{cmet}{A character which correspond to the clusterig method of the pRoloc package
knn, svm, naiveBayes, perTurbo, nnet (neural network) and rf (random forest).}

\item{train_size}{A numeric between 0 and 1 corresponding to train size in percentage}
}
\value{
a list containing the loss and the clustering score of the proteins which were miss located at each loop;
The mean loss and the mean clustering score of the proteins which were miss located
}
\description{
\code{ErrorClustering} perform n loop to calculate the loss of the clustering algorithm from the pRoloc package.
Take only the protein from known locations, separate randomly in 80/20 train set / test set.
Then count the number of mistake made by the algorithm.
}
\examples{
library(pRolocdata)
data(tan2009r1)
err_func(tan2009r1, 5)
}
\seealso{
\code{\link{svmOptimisation}} from pRoloc package and  \code{\link{datavisupca}} for more details
}
