% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FdynCon_n_Msn}
\alias{FdynCon_n_Msn}
\title{Data of dynamic organellar map, from Borner and al. 2016}
\format{
A MSnSet
}
\usage{
FdynCon_n_Msn

SdynCon_n_Msn

TdynCon_n_Msn

FdynEGF_n_Msn

SdynEGF_n_Msn

TdynEGF_n_Msn
}
\description{
A \code{\link{MSnSet}} containing the data from Borner and al. 2016.
 It contains 2235 proteins, with five fractions.
 F is for first replicate, S the second and T the third.
 Con is for control and EGF id for treated by EGF.
 n_Msn is because the data were normalized (see Borner and al. 2016 for more details)
 The best way to access its content is to use the functions :
 \code{\link{fData}} and \code{\link{exprs}},
 exprs show a tab of 5 columns (the five fractions) and 2235 rows (three replicates times two conditions)
 fData show a tab of 2 columns and 2235 rows containing the gene_name and the markers of each protein
 (their cellular location).
 Here \code{\link{pData}} isn't useful
}
\keyword{datasets}
