% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nainunknown.R
\name{nainunknown}
\alias{nainunknown}
\title{Fit the pRolocdata format}
\usage{
nainunknown(object, fcol = "markers")
}
\arguments{
\item{object}{A \code{\link{MSnSet}} object}

\item{fcol}{The name of the column where are the markers (organelles names)}
}
\value{
The object without NA values in the marker column of its  fData
}
\description{
\code{nainunknown} transform all NA values in the marker column of the fData of the object in "unknown".
}
\examples{

library(pRolocExtra)
nainunknown(tan2009r1)
}
