% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileInteract.R
\name{ProfileInteract}
\alias{ProfileInteract}
\title{Visualize interactive protein profile}
\usage{
ProfileInteract(
  obj,
  mrk = "markers",
  Organelle = "Golgi",
  Interact = TRUE,
  mytitle = FALSE,
  TITLE = "Profile",
  Clust = FALSE,
  one_pr = FALSE,
  protein = NULL,
  xLab = "Fractions",
  yLab = "Intensity"
)
}
\arguments{
\item{obj}{A \code{\link{MSnSet}} object}

\item{mrk}{The name of the column which contains the markers from the data}

\item{Organelle}{A character vector which contains the organelle(s) that you want to see the protein profile}

\item{Interact}{A logical argument to tell if you want an interactive plot}

\item{mytitle}{A logical argument to tell if you want to change the first part of the subtitle (name of the data)}

\item{TITLE}{A character which be your title (use with mytitle = TRUE)}

\item{Clust}{A logical argument to tell if you use clustered data.
(mrk should be equal to "knn" or orther clustering algorithm or any name of your data).}

\item{one_pr}{A logical argument to tell if you want to highlight a specific protein profile (or several)}

\item{protein}{A character vector which contains the proteins you want to highlight.
if NULL, show all the proteins}

\item{xLab}{A character which will be the name of the x-axis}

\item{yLab}{A character which will be the name of the y-axis}
}
\value{
A gpplot or ggplotly object : the protein profiles
}
\description{
\code{ProfileInteract} allows you to visualize protein profile from the organelle(s) you want
interactively or not, and highlight specific proteins.
It is based on the plotDist function from pRoloc package.
}
\examples{
library(pRolocdata)
data(tan2009r1)
Profil_interact(tan2009r1)
}
\seealso{
\code{\link{ggplotly}}, \code{\link{gghighlight}} and \code{\link{plotDist}} from Proloc package
}
