% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pRolocExtra.R
\docType{package}
\name{pRolocExtra}
\alias{pRolocExtra}
\title{pRolocExtra Package}
\description{
Contains several functions to generalize the pRoloc functions
to other data, to get better and more synthetic plot visualization.
}
\author{
Marc-Antoine GERAULT \email{marco.gerault@gmail.com}
}
