% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runpRolocExtra.R
\name{runpRolocExtra}
\alias{runpRolocExtra}
\title{Run the app to visualize all data from pRolocdata}
\usage{
runpRolocExtra()
}
\description{
\code{runpRolocExtra} works exactly the same as runExample from \code{\link{shiny}} package.
}
