% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyVisualization.R
\name{runShinyVisualization}
\alias{runShinyVisualization}
\title{Run the app to visualize all data from pRolocdata}
\usage{
runShinyVisualization()
}
\description{
\code{runShinyVisualization} works exactly the same as runExample from \code{\link{shiny}} package.
}
