% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgboostClassification.R
\name{xgboostClassification}
\alias{xgboostClassification}
\title{xgboost classification}
\usage{
xgboostClassification(
  object,
  assessRes,
  scores = c("prediction", "all", "none"),
  max_depth,
  gamma,
  nrounds = 1000,
  fcol = "markers",
  ...
)
}
\arguments{
\item{object}{An instance of class \code{"\linkS4class{MSnSet}"}.}

\item{assessRes}{An instance of class
\code{"\linkS4class{GenRegRes}"}, as generated by
\code{\link{xgboostOptimisation}}.}

\item{scores}{One of \code{"prediction"}, \code{"all"} or
\code{"none"} to report the score for the predicted class
only, for all classes or none.}

\item{max_depth}{If \code{assessRes} is missing, a \code{max_depth} must be
provided.}

\item{gamma}{If \code{assessRes} is missing, a \code{gamma} must
be provided.}

\item{fcol}{The feature meta-data containing marker definitions.
Default is \code{markers}.}

\item{...}{Additional parameters passed to \code{\link{xgb.train}} from
package \code{xgboost}.}
}
\value{
An instance of class \code{"\linkS4class{MSnSet}"} with
    \code{xgboost} and \code{xgboost.scores} feature variables storing the
    classification results and scores respectively.
}
\description{
Classification using the xgboost algorithm.
}
\examples{
library(pRolocExtra)
data(tan2009r1)
# reducing iterations
params <- xgboostOptimisation(tan2009r1, times = 3)
params
plot(params)
f1Count(params)
levelPlot(params)
getParams(params)
res <- xgboostClassification(tan2009r1, params)
getPredictions(res, fcol = "xgboost")
getPredictions(res, fcol = "xgboost", t = 0.75)
plot2D(res, fcol = "xgboost")
}
