% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxProfile.R
\name{BoxProfile}
\alias{BoxProfile}
\title{Box plot profile for dynamic organellar map}
\usage{
BoxProfile(
  obj,
  prot = "GRB2",
  Condition = "cond",
  mytit = TRUE,
  tit = "Box profile",
  cmet = "markers"
)
}
\arguments{
\item{obj}{A \code{\link{MSnSet}} object}

\item{prot}{A character vector which contains the protein(s) that you want to see the profiles between the two conditions}

\item{Condition}{The name of the column of your data which contains the condition for each protein (control or treated)}

\item{mytit}{Logical to tell if you want to print your own first element of the title}

\item{tit}{Character corresponding to the first element of the title, use when mytit = TRUE}

\item{cmet}{Character correspoonding to the column name containing the localization of each proteins}
}
\value{
a list containing the plot and the summary of the anova analysis
}
\description{
\code{BoxProfile} plot a box plot of specific proteins between two condition (control vs treated).
Perform also an anova analysis on the protein selected.
The box are only useful when there are replicates.
This allow to visualize quickly if the protein profile changed between the two conditions.
Use only when you have dataset with two different conditions.
}
