% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaletteWithoutGrey.R
\name{PaletteWithoutGrey}
\alias{PaletteWithoutGrey}
\title{Create a color palette}
\usage{
PaletteWithoutGrey(markers)
}
\arguments{
\item{markers}{A list containing all the markers of your data but it can be any list of character, factors, numeric, etc.}
}
\value{
A list of color which can be used for a plot
}
\description{
\code{PaletteWithoutGrey} create a color palette which not contains grey colors
 of the same length as the number of different markers of your \code{\link{MSnSet}} object.
}
